[domain] move MD5Hash into module
