+++ src/main/scala/net/kemitix/s3thorp/HashModified.scala
@@ -1,5 +1,7 @@
 package net.kemitix.s3thorp
 
+import net.kemitix.s3thorp.domain.MD5Hash
+
 final case class HashModified(hash: MD5Hash,
                               modified: LastModified) {
