+++ src/main/scala/net/kemitix/s3thorp/RemoteMetaData.scala
@@ -1,6 +1,6 @@
 package net.kemitix.s3thorp
 
-import net.kemitix.s3thorp.domain.RemoteKey
+import net.kemitix.s3thorp.domain.{MD5Hash, RemoteKey}
 
 final case class RemoteMetaData(remoteKey: RemoteKey,
                                 hash: MD5Hash,
