+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientMultiPartUploaderLogging.scala
@@ -1,7 +1,8 @@
 package net.kemitix.s3thorp.awssdk
 
 import com.amazonaws.services.s3.model.{AmazonS3Exception, InitiateMultipartUploadResult, UploadPartRequest, UploadPartResult}
-import net.kemitix.s3thorp.{Config, LocalFile, MD5Hash}
+import net.kemitix.s3thorp.domain.MD5Hash
+import net.kemitix.s3thorp.{Config, LocalFile}
 
 trait S3ClientMultiPartUploaderLogging
   extends S3ClientLogging {
