+++ src/test/scala/net/kemitix/s3thorp/ActionGeneratorSuite.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp
 
 import java.time.Instant
 
-import net.kemitix.s3thorp.domain.{Bucket, RemoteKey}
+import net.kemitix.s3thorp.domain.{Bucket, MD5Hash, RemoteKey}
 
 class ActionGeneratorSuite
   extends UnitTest
